Change getCategoriesWithTags() to getCategoriesWithEvents()
Checking for "does this category have events" results in a 22% improvement in loading time over "does this category have events with any tags". The previous functionality could be re-implemented in the future with clever queries or clever caching, but only if it doesn't increase the page load time disproportionately to the value of avoiding "aw, I clicked on a category in /tags and there weren't any tags for that category".