Remove migrations
This set of migrations reverses all of its own changes, so it makes more
sense to remove them and run all remaining migrations on a fresh copy
of the CakePHP 2 database. These migrations concerned the creation and
later removal of UTC time columns. Times will be stored in EDT for now.