Add AddTokenToUsers migration
Users.api_key is used to give a developer access to an API
in general, whereas Users.token is used to identify a user
as the person authorized to update the database. So in general,
one API key would be used for an application but every user
using it would use their own personal access token.